[staccmr-39] added initial setup
